// src/models/apflora/Apberuebersicht.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ProjektId } from './Projekt';

/** Identifier type for apflora.apberuebersicht */
export type ApberuebersichtId = string & { __brand: 'apflora.apberuebersicht' };

/** Represents the table apflora.apberuebersicht */
export default interface Apberuebersicht {
  /** Primärschlüssel */
  id: ApberuebersichtId;

  /** Berichtsjahr. Zusammen mit proj_id eindeutig */
  jahr: number | null;

  /** Bemerkungen zur Artübersicht */
  bemerkungen: string | null;

  /** Von wem wurde der Datensatz zuletzt geändert? */
  changed_by: string | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "projekt" */
  proj_id: ProjektId | null;

  /** Datum, an dem die Daten von AP, Pop und TPop historisiert wurden */
  history_date: Date | null;

  created_at: Date;

  updated_at: Date;

  history_fixed: boolean | null;
}

/** Represents the initializer for the table apflora.apberuebersicht */
export interface ApberuebersichtInitializer {
  /**
   * Primärschlüssel
   * Default value: uuid_generate_v1mc()
   */
  id?: ApberuebersichtId;

  /** Berichtsjahr. Zusammen mit proj_id eindeutig */
  jahr?: number | null;

  /** Bemerkungen zur Artübersicht */
  bemerkungen?: string | null;

  /**
   * Von wem wurde der Datensatz zuletzt geändert?
   * Default value: NULL::character varying
   */
  changed_by?: string | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "projekt" */
  proj_id?: ProjektId | null;

  /** Datum, an dem die Daten von AP, Pop und TPop historisiert wurden */
  history_date?: Date | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  /** Default value: false */
  history_fixed?: boolean | null;
}

/** Represents the mutator for the table apflora.apberuebersicht */
export interface ApberuebersichtMutator {
  /** Primärschlüssel */
  id?: ApberuebersichtId;

  /** Berichtsjahr. Zusammen mit proj_id eindeutig */
  jahr?: number | null;

  /** Bemerkungen zur Artübersicht */
  bemerkungen?: string | null;

  /** Von wem wurde der Datensatz zuletzt geändert? */
  changed_by?: string | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "projekt" */
  proj_id?: ProjektId | null;

  /** Datum, an dem die Daten von AP, Pop und TPop historisiert wurden */
  history_date?: Date | null;

  created_at?: Date;

  updated_at?: Date;

  history_fixed?: boolean | null;
}